Move InputFile into a dedicated header.
